requirements.txt
pytest
streamlit
langchain-ollama